New primitive for forall: "mincount" existence checks.
Rather than the (rather fragile) old mechanism of a magic "forall"
relation, which turned out to be a bottleneck in some cases (just
populating the magic table caused an OOM!), this change adds a new
condition to Scan ops in the RAM program that requires an
existence-check scan to match a minimum number of records. This can be
used in conjunction with a count aggregator over the domain in order to
trigger a rule when all elements in the domain have been added to a
particular relation. This change removes the old forall mechanism as
well.

Example use case:

    .decl Domain(?k1:symbol, ?k2:symbol, ?v1:symbol, ?v2:symbol)
    .decl Input(?k1:symbol, ?k2:symbol, ?v1:symbol, ?v2:symbol)

    .decl DomainCount(?k1:symbol, ?k2:symbol, ?count:number)
    DomainCount(?k1, ?k2, ?count) :-
      Domain(?k1, ?k2, _, _),
      ?count = count : { Domain(?k1, ?k2, ?v1, ?v2) }.

    .decl FilteredInput(?k1:symbol, ?k2:symbol, ?v1:symbol, ?v2:symbol)
    FilteredInput(?k1, ?k2, ?v1, ?v2) :-
      Input(?k1, ?k2, ?v1, ?v2),
      Domain(?k1, ?k2, ?v1, ?v2).

    .decl All(?k1:symbol, ?k2:symbol)
    All(?k1, ?k2) :-
      DomainCount(?k1, ?k2, ?dcount),
      mincount(FilteredInput(?k1, ?k2, _, _), ?dcount).